+++ scripts/lint-allow-test-rule-refs.allowlist.json
@@ -77,13 +77,9 @@
   "tests/intel.test.cjs :: source-text-is-the-product",
   "tests/inventory-headings-countfree.test.cjs :: source-text-is-the-product",
   "tests/ios-scaffold-safety.test.cjs :: source-text-is-the-product",
-  "tests/issue-2639-codex-toml-neutralization.test.cjs :: source-text-is-the-product",
+  "tests/legacy-cleanup.test.cjs :: integration-test-input",
   "tests/issue-429-comment-text-gate.test.cjs :: source-text-is-the-product",
   "tests/issue-498-update-backup-runtime-dir.test.cjs :: source-text-is-the-product",
-  "tests/issue-57-runtime-install-no-drift.test.cjs :: structural-regression-guard",
-  "tests/issue-607-installer-dry-run.install.test.cjs :: integration-test-input",
-  "tests/issue-607-legacy-cleanup.test.cjs :: integration-test-input",
-  "tests/issue-787-cline-hooks-agents.test.cjs :: source-text-is-the-product",
   "tests/issue-815-update-next-channel.test.cjs :: source-text-is-the-product",
   "tests/locking-bugs-1909-1916-1925-1927.test.cjs :: architectural-invariant",
   "tests/mcp-tool-inheritance.test.cjs :: source-text-is-the-product",
