+++ scripts/lint-allow-test-rule-refs.ceiling.json
@@ -1,4 +1,4 @@
 {
-  "maxFiles": 309,
+  "maxFiles": 305,
   "grace": 3
 }
