+++ tests/legacy-cleanup.test.cjs
