+++ tests/runtime-artifact-layout-surface.test.cjs
@@ -1415,3 +1415,114 @@ describe('installOpencodeFamilySkills destination parity (#2911 sibling coverage
     });
   }
 });
+
+// #69 regression, folded from issue-69-surface-keeps-nested.test.cjs:
+// stageSkillsForRuntimeAsSkills gated nesting on `resolvedProfile.skills === '*'`
+// (the sentinel). applySurface → resolveSurface materializes the full profile
+// into a concrete Set<string>, so the sentinel check was never true on the
+// surface path, causing applySurface to re-flatten a nested install. Fix
+// (install-profiles.cts): gate nesting on full OR full-equivalent (all
+// routerStems present in the concrete Set).
+//
+// #924: Claude was reverted to FLAT, so the claude case below asserts the
+// flat layout is preserved (not re-nested) rather than a nested one.
+describe('issue-69: applySurface preserves nested skill layout (no re-flatten)', () => {
+  test('cline global full: applySurface keeps 6 router dirs and nested gsd-ns-manage/skills/help/SKILL.md', (t) => {
+    process.env.GSD_TEST_MODE = '1';
+    const { installRuntimeArtifacts } = require('../gsd-core/bin/lib/install-engine.cjs');
+    const dir = tmpDir('gsd-69-surface-');
+    t.after(() => { try { cleanup(dir); } catch { /* best-effort */ } });
+
+    // Step 1: full install
+    const manifest = loadSkillsManifest(REAL_COMMANDS_DIR);
+    const resolved = resolveProfile({ modes: ['full'], manifest });
+    installRuntimeArtifacts('cline', dir, 'global', resolved);
+
+    const skillsDir = path.join(dir, 'skills');
+
+    // Sanity: install must produce nested layout (6 top-level router dirs)
+    const topLevelAfterInstall = fs.readdirSync(skillsDir).filter((n) => n.startsWith('gsd-'));
+    assert.strictEqual(
+      topLevelAfterInstall.length,
+      6,
+      `Install must produce exactly 6 gsd-* top-level dirs (routers). Got ${topLevelAfterInstall.length}: [${topLevelAfterInstall.join(', ')}]`,
+    );
+    assert.ok(
+      fs.existsSync(path.join(skillsDir, 'gsd-ns-workflow', 'skills', 'plan-phase', 'SKILL.md')),
+      'After install: gsd-ns-workflow/skills/plan-phase/SKILL.md must exist',
+    );
+
+    // Step 2: applySurface (full surface, no surface state file → resolves to full)
+    const layout = resolveRuntimeArtifactLayout('cline', dir, 'global');
+    applySurface(dir, layout, manifest);
+
+    // Step 3: assert nested layout is preserved after applySurface
+    const topLevelAfterSurface = fs.readdirSync(skillsDir).filter((n) => n.startsWith('gsd-'));
+    assert.strictEqual(
+      topLevelAfterSurface.length,
+      6,
+      `After applySurface: expected exactly 6 gsd-* top-level dirs (routers only). Got ${topLevelAfterSurface.length}: [${topLevelAfterSurface.join(', ')}]. ` +
+      'Re-flattening detected: applySurface must preserve nested layout (#69 regression).',
+    );
+
+    // The nested SKILL.md must still exist (not re-flattened to top-level concrete dir)
+    assert.ok(
+      fs.existsSync(path.join(skillsDir, 'gsd-ns-workflow', 'skills', 'plan-phase', 'SKILL.md')),
+      'After applySurface: gsd-ns-workflow/skills/plan-phase/SKILL.md must still exist (nested layout preserved)',
+    );
+
+    // The concrete skill must NOT have been promoted to a top-level flat dir
+    assert.ok(
+      !fs.existsSync(path.join(skillsDir, 'gsd-plan-phase', 'SKILL.md')),
+      'After applySurface: gsd-plan-phase/ must NOT exist at top level (#69 re-flatten regression guard)',
+    );
+  });
+
+  // #924 companion: Claude must use FLAT layout and applySurface must NOT re-nest it.
+  test('claude global full: install produces flat layout and applySurface preserves it (#924)', (t) => {
+    process.env.GSD_TEST_MODE = '1';
+    const { installRuntimeArtifacts } = require('../gsd-core/bin/lib/install-engine.cjs');
+    const dir = tmpDir('gsd-924-69-');
+    t.after(() => { try { cleanup(dir); } catch { /* best-effort */ } });
+
+    const manifest = loadSkillsManifest(REAL_COMMANDS_DIR);
+    const resolved = resolveProfile({ modes: ['full'], manifest });
+    installRuntimeArtifacts('claude', dir, 'global', resolved);
+
+    const skillsDir = path.join(dir, 'skills');
+
+    // Install must produce FLAT layout (>= 60 gsd-* dirs)
+    const topLevelAfterInstall = fs.readdirSync(skillsDir).filter((n) => n.startsWith('gsd-'));
+    assert.ok(
+      topLevelAfterInstall.length >= 60,
+      `Claude install must produce >= 60 gsd-* top-level dirs (flat, #924). Got ${topLevelAfterInstall.length}.`,
+    );
+
+    // gsd-plan-phase must be directly at top level
+    assert.ok(
+      fs.existsSync(path.join(skillsDir, 'gsd-plan-phase', 'SKILL.md')),
+      'After claude install: gsd-plan-phase/SKILL.md must be at top level (flat layout, #924)',
+    );
+
+    // No nested skills/ subdirs under gsd-ns-* in Claude
+    assert.ok(
+      !fs.existsSync(path.join(skillsDir, 'gsd-ns-workflow', 'skills')),
+      'After claude install: gsd-ns-workflow/skills/ must NOT exist (flat layout, no nesting, #924)',
+    );
+
+    // applySurface must preserve flat layout
+    const layout = resolveRuntimeArtifactLayout('claude', dir, 'global');
+    applySurface(dir, layout, manifest);
+
+    const topLevelAfterSurface = fs.readdirSync(skillsDir).filter((n) => n.startsWith('gsd-'));
+    assert.ok(
+      topLevelAfterSurface.length >= 60,
+      `After applySurface: claude must still have >= 60 gsd-* dirs (flat preserved). Got ${topLevelAfterSurface.length}.`,
+    );
+
+    assert.ok(
+      fs.existsSync(path.join(skillsDir, 'gsd-plan-phase', 'SKILL.md')),
+      'After applySurface: gsd-plan-phase/SKILL.md must remain at top level (#924)',
+    );
+  });
+});
